apify-actors/Developer_Tools/domain-inspector/requirements.txt
apify < 4.0.0
dnspython
python-whois
httpx